+++ build.sh
@@ -16,7 +16,7 @@ set -e
 # You can override the version of the core library
 # Otherwise, use the default value
 if [ -z "$REALM_CORE_VERSION" ]; then
-    REALM_CORE_VERSION=0.83.0
+    REALM_CORE_VERSION=0.83.1
 fi
 
 PATH=/usr/local/bin:/usr/bin:/bin:/usr/libexec:$PATH