+++ CHANGELOG.md
@@ -7,9 +7,11 @@
 
 ### Enhancements
 
+* Add support for sorting RLMArray properties.
 * Speed up inserting objects with `addObject:` by ~20%.
 * `readonly` properties are automatically ignored rather than having to be
   added to `ignoredProperties`.
+* Updating to core library version 0.83.1.
 
 ### Bugfixes